wip: start refactoring
