+++ .editorconfig
@@ -26,7 +26,7 @@ ij_java_align_multiline_annotation_parameters = false
 ij_java_align_multiline_array_initializer_expression = false
 ij_java_align_multiline_assignment = false
 ij_java_align_multiline_binary_operation = false
-ij_java_align_multiline_chained_methods = false
+ij_java_align_multiline_chained_methods = true
 ij_java_align_multiline_deconstruction_list_components = true
 ij_java_align_multiline_extends_list = false
 ij_java_align_multiline_for = false
