--- lib/net/lecigne/somafm-recentlib/0.2.2/_remote.repositories
@@ -1,4 +1,0 @@
-#NOTE: This is a Maven Resolver internal implementation file, its format can be changed without prior notice.
-#Sun Oct 26 12:30:12 CET 2025
-somafm-recentlib-0.2.2.jar>=
-somafm-recentlib-0.2.2.pom>=
