--- lib/net/lecigne/somafm-recentlib/maven-metadata-local.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<metadata>
-  <groupId>net.lecigne</groupId>
-  <artifactId>somafm-recentlib</artifactId>
-  <versioning>
-    <release>0.2.2</release>
-    <versions>
-      <version>0.2.2</version>
-    </versions>
-    <lastUpdated>20251026113012</lastUpdated>
-  </versioning>
-</metadata>
