--- src/main/java/net/lecigne/somafm/domain/UnknownChannelException.java
@@ -1,9 +1,0 @@
-package net.lecigne.somafm.domain;
-
-public class UnknownChannelException extends RuntimeException {
-
-  public UnknownChannelException(String channel) {
-    super("Unknown channel: " + channel);
-  }
-
-}
