+++ src/main/java/net/lecigne/somafm/history/adapters/in/cli/DisplayedBroadcastMapper.java
@@ -1,0 +1,25 @@
+package net.lecigne.somafm.history.adapters.in.cli;
+
+import java.time.LocalDateTime;
+import java.time.ZoneId;
+import java.time.format.DateTimeFormatter;
+import lombok.RequiredArgsConstructor;
+import net.lecigne.somafm.history.domain.DisplayableBroadcast;
+import net.lecigne.somafm.recentlib.Broadcast;
+
+@RequiredArgsConstructor
+class DisplayedBroadcastMapper {
+
+  private final ZoneId zoneId;
+  private static final DateTimeFormatter FORMATTER = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss");
+
+  DisplayableBroadcast map(Broadcast broadcast) {
+    return DisplayableBroadcast.builder()
+        .time(LocalDateTime.ofInstant(broadcast.time(), zoneId).format(FORMATTER))
+        .channel(broadcast.channel().publicName())
+        .artist(broadcast.song().artist() != null ? broadcast.song().artist().name() : null)
+        .title(broadcast.song().title())
+        .build();
+  }
+
+}
