+++ src/main/java/net/lecigne/somafm/history/application/ports/in/SomaFmSongHistory.java
@@ -1,0 +1,9 @@
+package net.lecigne.somafm.history.application.ports.in;
+
+import java.util.List;
+import net.lecigne.somafm.history.domain.SomaFmCommand;
+import net.lecigne.somafm.recentlib.Broadcast;
+
+public interface SomaFmSongHistory {
+  List<Broadcast> run(SomaFmCommand command);
+}
