+++ src/main/java/net/lecigne/somafm/history/bootstrap/config/SomaFmConfig.java
@@ -1,0 +1,35 @@
+package net.lecigne.somafm.history.bootstrap.config;
+
+import com.typesafe.config.Optional;
+import java.util.Objects;
+import java.util.stream.Stream;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@NoArgsConstructor
+@Getter
+@Setter
+public class SomaFmConfig {
+
+  public static final String ROOT_CONFIG = "config";
+
+  private String userAgent;
+  private String timezone;
+  @Optional
+  private DbConfig db;
+
+  @NoArgsConstructor
+  @Getter
+  @Setter
+  public static class DbConfig {
+    private String url;
+    private String user;
+    private String password;
+  }
+
+  public boolean isDbActivated() {
+    return Objects.nonNull(db) && Stream.of(db.getUrl(), db.getUser(), db.getPassword()).allMatch(Objects::nonNull);
+  }
+
+}
