+++ src/main/java/net/lecigne/somafm/history/domain/Action.java
@@ -1,0 +1,21 @@
+package net.lecigne.somafm.history.domain;
+
+import java.util.Arrays;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+@Getter
+public enum Action {
+  DISPLAY("display"), SAVE("save");
+
+  private final String actionName;
+
+  public static Action getValue(String actionName) {
+    return Arrays.stream(Action.values())
+        .filter(value -> value.actionName.equals(actionName))
+        .findFirst()
+        .orElse(DISPLAY);
+  }
+
+}
