+++ src/main/java/net/lecigne/somafm/history/domain/SomaFmCommand.java
@@ -1,0 +1,6 @@
+package net.lecigne.somafm.history.domain;
+
+import net.lecigne.somafm.recentlib.Channel;
+
+public record SomaFmCommand(Action action, Channel channel) {
+}
