+++ src/main/java/net/lecigne/somafm/history/domain/UnknownChannelException.java
@@ -1,0 +1,9 @@
+package net.lecigne.somafm.history.domain;
+
+public class UnknownChannelException extends RuntimeException {
+
+  public UnknownChannelException(String channel) {
+    super("Unknown channel: " + channel);
+  }
+
+}
