+++ src/main/resources/db/migration/V3__normalize_broadcasts_keys.sql
@@ -1,0 +1,18 @@
+ALTER TABLE broadcasts
+ADD COLUMN id BIGINT GENERATED BY DEFAULT AS IDENTITY;
+
+UPDATE broadcasts
+SET id = DEFAULT
+WHERE id IS NULL;
+
+ALTER TABLE broadcasts
+ALTER COLUMN id SET NOT NULL;
+
+ALTER TABLE broadcasts
+DROP CONSTRAINT broadcasts_pkey;
+
+ALTER TABLE broadcasts
+ADD CONSTRAINT broadcasts_pkey PRIMARY KEY (id);
+
+ALTER TABLE broadcasts
+ADD CONSTRAINT broadcasts_utc_time_channel_key UNIQUE (utc_time, channel);
