+++ src/main/resources/logback.xml
@@ -5,7 +5,7 @@
     </encoder>
   </appender>
   <root level="OFF"/>
-  <logger name="net.lecigne.somafm" level="INFO">
+  <logger name="net.lecigne.somafm.history" level="INFO">
     <appender-ref ref="CONSOLE"/>
   </logger>
 </configuration>
