--- src/test/resources/data/dronezone_no_table.html
@@ -1,3 +1,0 @@
-<div id="playinc">
-  <p>Hello!</p>
-</div>
